+++ R/RcppExports.R
@@ -279,6 +279,10 @@ carbon_carbonCompartments <- function(x, biomassUnits = "g_m2") {
     .Call(`_medfate_carbonCompartments`, x, biomassUnits)
 }
 
+.addSPWBCommunicationStructures <- function(x) {
+    invisible(.Call(`_medfate_addSPWBCommunicationStructures`, x))
+}
+
 .criticalFirelineIntensity <- function(CBH, M) {
     .Call(`_medfate_criticalFirelineIntensity`, CBH, M)
 }