+++ src/communication_structures.h
@@ -1,0 +1,10 @@
+#include <Rcpp.h>
+
+#ifndef COMMUNICATION_STRUCTURES_H
+#define COMMUNICATION_STRUCTURES_H
+#endif
+using namespace Rcpp;
+
+void addSPWBCommunicationStructures(List x);
+
+List internalLongWaveRadiation(int ncanlayers);